+++ scripts/calculate_ims.py
@@ -7,93 +7,7 @@ from typing import List, Optional
 from IM import ims, waveform_reading
 
 
-class IM(StrEnum):
-    PGA = "PGA"
-    PGV = "PGV"
-    pSA = "pSA"
-    CAV = "CAV"
-    CAV5 = "CAV5"
-    MMI = "MMI"
-    Ds575 = "Ds575"
-    Ds595 = "Ds595"
-    AI = "AI"
-    FAS = "FAS"
-
 app = typer.Typer()
-
-
-def calculate_ims(
-    waveform: np.ndarray,
-    dt: float,
-    ims_list: List[IM],
-    components: List[str],
-    periods: Optional[List[float]] = None,
-    frequencies: Optional[List[float]] = typer.Option(None, help="Frequencies required for FAS"),
-):
-    """
-    Calculate intensity measures for a waveform.
-
-    Parameters
-    ----------
-    waveform : np.ndarray
-        Waveform data as a NumPy array.
-    ims_list : list of IM
-        List of intensity measures (IMs) to calculate, e.g., ['PGA', 'pSA', 'CAV'].
-    components : list of str
-        List of waveform components to process, e.g., ['X', 'Y', 'Z'].
-    periods : list of float, optional
-        List of periods required for calculating the pseudo-spectral acceleration (pSA).
-    frequencies : list of float, optional
-        List of frequencies required for calculating the Fourier amplitude spectrum (FAS).
-
-    Returns
-    -------
-    None
-        Displays the calculated intensity measures in the terminal or command output.
-    """
-    # Validate components
-    if not components:
-        typer.echo("Components list cannot be empty.")
-        raise typer.Exit()
-
-    # Placeholder for results
-    results = {}
-
-    # Iterate through IMs and calculate them
-    for im in ims_list:
-        typer.echo(f"Calculating {im}...")
-        try:
-            if im == IM.PGA:
-                results[im] = ims.peak_ground_acceleration(waveform)
-            elif im == IM.PGV:
-                results[im] = ims.calculate_pgv(waveform, components)
-            elif im == IM.pSA:
-                if not periods:
-                    typer.echo("Periods are required for pSA.")
-                    raise typer.Exit()
-                results[im] = ims.calculate_psa(waveform, components, periods)
-            elif im == IM.CAV:
-                results[im] = ims.calculate_cav(waveform, components)
-            elif im == IM.CAV5:
-                results[im] = ims.calculate_cav5(waveform, components)
-            elif im == IM.MMI:
-                results[im] = ims.calculate_mmi(waveform, components)
-            elif im == IM.Ds575:
-                results[im] = ims.calculate_ds575(waveform, components)
-            elif im == IM.Ds595:
-                results[im] = ims.calculate_ds595(waveform, components)
-            elif im == IM.AI:
-                results[im] = ims.calculate_ai(waveform, components)
-            elif im == IM.FAS:
-                if not frequencies:
-                    typer.echo("Frequencies are required for FAS.")
-                    raise typer.Exit()
-                results[im] = ims.calculate_fas(waveform, components, frequencies)
-            else:
-                typer.echo(f"Unsupported IM: {im}")
-        except Exception as e:
-            typer.echo(f"Error calculating {im}: {e}")
-            raise typer.Exit()
 
 
 @app.command()
